issue with search by type of payment
