+++ src/components/templates/AdministratorPaymentsTemplate/AdministratorPaymentsTemplate.styles.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const Center = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 70vh;
+`;
