+++ src/helpers/paymentHelpers/paymenttypeAndMethodConst.js
@@ -1,9 +1,11 @@
 export const typeOfPayment = [
+  '',
   'Opłata adopcyjna',
   'Jednorazowy przelew',
   'Wirtualny opiekun-opłata cykliczna',
 ];
 export const paymentMethod = [
+  '',
   'Karta płatnicza',
   'Blik',
   'Przelew bankowy',
